Fix function names
The function names appended with `T` in the `Gameplay` module didn't
make sense. I incorrectly assumed the appended `T` in functions like
`execStateT` were there because of some convention around transformers,
but it's actually the literal `StateT` that's appended.

Change the method names to unfollow that wrong convention and change
them to more descriptive names.